+++ Command/ExportCommand.php
@@ -16,11 +16,6 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
-/**
- * Class ExportCommand
- *
- * @package Semaio\ConfigImportExport\Command
- */
 class ExportCommand extends AbstractCommand
 {
     /**