+++ Model/File/FinderInterface.php
@@ -6,11 +6,6 @@
 
 namespace Semaio\ConfigImportExport\Model\File;
 
-/**
- * Interface FinderInterface
- *
- * @package Semaio\ConfigImportExport\Model\File
- */
 interface FinderInterface
 {
     /**
@@ -37,4 +32,9 @@ interface FinderInterface
      * @param mixed $format
      */
     public function setFormat($format);
+
+    /**
+     * @param mixed $depth
+     */
+    public function setDepth($depth);
 }