+++ Model/File/Reader/ReaderInterface.php
@@ -6,11 +6,6 @@
 
 namespace Semaio\ConfigImportExport\Model\File\Reader;
 
-/**
- * Interface ReaderInterface
- *
- * @package Semaio\ConfigImportExport\Model\File\Reader
- */
 interface ReaderInterface
 {
     /**