+++ Model/File/Reader/YamlReader.php
@@ -8,11 +8,6 @@ namespace Semaio\ConfigImportExport\Model\File\Reader;
 
 use Symfony\Component\Yaml\Yaml as SymfonyYaml;
 
-/**
- * Class YamlReader
- *
- * @package Semaio\ConfigImportExport\Model\File\Reader
- */
 class YamlReader extends AbstractReader
 {
     /**